+++ properties.schema
@@ -18,7 +18,7 @@
             "type": "string",
             "required": true,
             "enum": ["full-width", "half-width", "both"],
-            "default": "half-width",
+            "default": "full-width",
             "editorOnly": true
         },
         "instruction": {
